+++ packages/basehub/src/common-types.ts
@@ -1,0 +1,14 @@
+export type ResolvedRef = { ref: string } & (
+  | {
+      type: "commit";
+      id: string;
+      message: string;
+    }
+  | {
+      type: "branch";
+      id: string;
+      name: string;
+      git?: { branch?: string | null };
+      createSuggestedBranchLink?: string;
+    }
+);