+++ playground/src/lib/constants.ts
@@ -1,4 +1,0 @@
-export const siteURL = new URL(
-  process.env.NEXT_PUBLIC_SITE_URL as string | URL
-);
-export const siteOrigin = siteURL.origin;