+++ playground/CHANGELOG.md
@@ -1,5 +1,166 @@
 # playground
 
+## 0.0.148
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.6
+
+## 0.0.147
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.5
+
+## 0.0.146
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.4
+
+## 0.0.145
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.3
+
+## 0.0.144
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.2
+
+## 0.0.143
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.1
+
+## 0.0.142
+
+### Patch Changes
+
+- Updated dependencies [8945431]
+- Updated dependencies [8945431]
+- Updated dependencies [dbe9b38]
+- Updated dependencies [1f74c57]
+- Updated dependencies [8945431]
+- Updated dependencies [dbe9b38]
+- Updated dependencies [b5a5bd6]
+- Updated dependencies [9f17428]
+- Updated dependencies [8d2af40]
+- Updated dependencies [1f24ced]
+- Updated dependencies [7d76da3]
+- Updated dependencies [bc9d438]
+- Updated dependencies [e41cdd0]
+- Updated dependencies [7d76da3]
+  - basehub@7.5.0
+
+## 0.0.142-canary.12
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.12
+
+## 0.0.142-canary.11
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.11
+
+## 0.0.142-canary.10
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.10
+
+## 0.0.142-canary.9
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.9
+
+## 0.0.142-canary.8
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.8
+
+## 0.0.142-canary.7
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.7
+
+## 0.0.142-canary.6
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - basehub@7.5.0-canary.6
+
+## 0.0.142-canary.5
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.5
+
+## 0.0.142-canary.4
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.4
+
+## 0.0.142-canary.3
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.3
+
+## 0.0.142-canary.2
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.2
+
+## 0.0.142-canary.1
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.1
+
+## 0.0.142-canary.0
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.5.0-canary.0
+
+## 0.0.141
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.4.2
+
 ## 0.0.140
 
 ### Patch Changes